microservices/math/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycalc-tigxy @ git+https://github.com/tigxy/pycalc.git@7ab71a8d249d3afd864b88809ab6ab62c4b23433
requests==2.26.0
urllib3==1.26.7
Werkzeug==2.0.2
wincertstore==0.2
